Moved CSEXRB functionality to parent class ossimNitfGenericTag
