+++ include/ossim/support_data/ossimNitfCsexrbTag.h
@@ -1,37 +1,12 @@
-//---
-// File: ossimNitfCsexrbTag.h
-//---
+//
+// Created by Ryan Feldbush on 6/18/25.
+//
 
-#include <ossim/support_data/ossimNitfRegisteredTag.h>
-#include <vector>
-#include <map>
+#include <ossim/support_data/ossimNitfGenericTag.h>
+
 #ifndef ossimNitfCsexrbTag_HEADER
 #define ossimNitfCsexrbTag_HEADER 1
 
-/**
- * @class ossimNitfCsexrbTag
- */
-class OSSIM_DLL ossimNitfCsexrbTag : public ossimNitfRegisteredTag
-{
-public:
-   ossimNitfCsexrbTag();
-
-   virtual void parseStream(std::istream& in);
-
-   virtual void writeStream(std::ostream& out);
-
-   virtual void clearFields();
-
-   virtual std::ostream& print(std::ostream& out,
-                               const std::string& prefix) const;
-
-   ossimString get(ossimString fieldName);
-
-protected:
-
-   std::map<ossimString, ossimString> m_fields_map;
-   std::vector<std::pair<ossimString, ossimString>> m_fields_vector;
-
+class OSSIM_DLL ossimNitfCsexrbTag : public ossimNitfGenericTag {
 };
-
 #endif
